Tons of changes. The biggest are moving and resizing windows.
1. Refactored some code so that interface methods are "exported". This
makes dealing with namespace collisions a little more elegant and makes
it easier to identify which methods are "interface" methods. Not really
a valid use of "exporting", but a window manager will never be used as
an importable package...

2. A basic model of window stacking and focus has been implemented.

3. Moving and resizing using the new "mouse dragging" feature in xgbutil
seems to work okay. (I've ported my code nearly verbatim from pyndow for
the moving/resizing guts---it's quite complex.)