Things are really starting to come together now. Much of the layout and workspace related logic is working now. i.e., tiling, untiling, moving windows to different heads, maximizing, etc.
What's surprising is that the crux of the consistency behind maintaining
window state is the combined effect of "saveGeom" and "loadGeom."

They were initially intended for simplistic use---to save the geometry
of the window before maximization so that it could be restored to its
original position.

It now encapsulates a lot more information than just the geometry
(perhaps a name change is in order). Namely:

1) The x/y/width/height traditional geometry.
2) Whether it is in a maximized state or not.
3) The kind of frame being used.
4) The current head geometry.

All of this information allows seamless movement between different
states. For example, let's say you tile a maximized window. Upon
untiling, the window should go back to its maximized state. But what if
the window is restored after this?

Since saveGeom and loadGeom can maintain arbitrary geometrical states,
this is as simple as having an "unmaximized" and a "before_tiling"
state.

This mechanism is also extremely useful when showing/hiding workspaces.

The coolest thing is that it totally encapsulates the logic of "this
client wasn't on this head before, so let's translate it geometry into
the new head."

Finally, the *geometry* of the head is captured rather than the head
number because the heads can change. (i.e., what was a valid head could
become invalid.)