Corrects workspace geometries on ApplyStruts
Fixes incorrect workspace geometries when Reload'ing.
This fixes a problem with incorrect window handling after changing the resolution or display layout.
One easy way to reproduce this was to change the resolution, then maximize a window (Mod4-x). This would result in the window being resized to the old, now changed size.