Fix a bug in synchronous command execution.
The point of syncRun was to force synchronous execution of a command
with respect to the X event loop. While it achieved this, it didn't
actually enforce synchronous execution with respect to *other* commands.
Namely, multiple invocations of syncRun can happen in parallel, which
can result in the commands being interleaved.

We force the issue by ensuring only one invocation of syncRun can execute
at any point in time. It's not pretty, but it works.