focus: fix data race in focus package
The focus package was maintaining a global map (and slice) of tracked
clients, but wasn't using synchronization to access it. We fix that
here.

We also used to export 'Clients' directly, but of course, this leaves it
unprotected. We unexport it and provide a 'Clients()' function instead,
which does proper synchronization.

Unfortunately, it looks like the actual clients returned are themselves
not goroutine safe, so this is really just a small bandaide. Sigh.